fix product search function

+++ md5_phone_store_management/src/main/java/com/example/md5_phone_store_management/service/IProductService.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface IProductService {
     //Tuấn Anh
     Page<Product> findAll(Pageable pageable);
-    Page<Product> searchProductByNameAndSupplier_NameAndPurchasePrice(String name, String supplierName, double purchasePrice, Pageable pageable);
+    Page<Product> searchProductByNameAndSupplier_NameAndPurchasePrice(String name, String supplierName, int purchasePrice, Pageable pageable);
 
     //Đình Anh
     void saveProduct(Product product);
